// src/generated/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
};
// ...
export type Bonus = {
  __typename?: 'Bonus';
  asset: Scalars['String']['output'];
  /** How much activations must be for new level */
  condition?: Maybe<Scalars['Float']['output']>;
  count: Scalars['Float']['output'];
  createdAt: Scalars['DateTime']['output'];
  description: Scalars['String']['output'];
  id: Scalars['String']['output'];
  isActive: Scalars['Boolean']['output'];
  level?: Maybe<BonusLevelType>;
  payload?: Maybe<Scalars['JSON']['output']>;
  updatedAt: Scalars['DateTime']['output'];
  valueType: BonusValueType;
};

export enum BonusLevelType {
  Junior = 'JUNIOR',
  Middle = 'MIDDLE',
  Senior = 'SENIOR'
}

export type BonusTicket = {
  __typename?: 'BonusTicket';
  activeTill?: Maybe<Scalars['DateTime']['output']>;
  code: Scalars['String']['output'];
  createdAt: Scalars['DateTime']['output'];
  id: Scalars['String']['output'];
  isActive: Scalars['Boolean']['output'];
  ticketType: BonusTicketType;
  updatedAt: Scalars['DateTime']['output'];
};

export enum BonusTicketType {
  Const = 'CONST',
  Disposable = 'DISPOSABLE'
}

export type BonusTicketWholeType = {
  __typename?: 'BonusTicketWholeType';
  activeTill?: Maybe<Scalars['DateTime']['output']>;
  bonus?: Maybe<Bonus>;
  code: Scalars['String']['output'];
  createdAt: Scalars['DateTime']['output'];
  id: Scalars['String']['output'];
  isActive: Scalars['Boolean']['output'];
  ticketType: BonusTicketType;
  updatedAt: Scalars['DateTime']['output'];
  user?: Maybe<User>;
};

export enum BonusValueType {
  Percent = 'PERCENT',
  Point = 'POINT'
}
// ...
export type User = {
  __typename?: 'User';
  client_assessment?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['DateTime']['output'];
  email: Scalars['String']['output'];
  enable_notifications: Scalars['Boolean']['output'];
  google_id?: Maybe<Scalars['String']['output']>;
  id: Scalars['String']['output'];
  instagram?: Maybe<Scalars['String']['output']>;
  isFullAuth: Scalars['Boolean']['output'];
  isGoogleAuth?: Maybe<Scalars['Boolean']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  password?: Maybe<Scalars['String']['output']>;
  phone?: Maybe<Scalars['String']['output']>;
  picture?: Maybe<Scalars['String']['output']>;
  rate?: Maybe<Scalars['Float']['output']>;
  role: UserRole;
  telegram?: Maybe<Scalars['String']['output']>;
  updatedAt: Scalars['DateTime']['output'];
  verified_email?: Maybe<Scalars['Boolean']['output']>;
};

export enum UserRole {
  Admin = 'admin',
  Analyt = 'analyt',
  Blocked = 'blocked',
  Manager = 'manager',
  SuperManager = 'super_manager',
  User = 'user'
}